+++ cmd/pg2sink/cmd/version.go
@@ -1,0 +1,21 @@
+package cmd
+
+import (
+	"fmt"
+	"runtime"
+)
+
+// Set by goreleaser
+var (
+	Version   = "dev"
+	Commit    = "none"
+	Date      = "unknown"
+	GoVersion = runtime.Version()
+)
+
+func versionStanza() string {
+	return fmt.Sprintf(
+		"pg2sink Version: %v\nGit SHA: %v\nGo Version: %v\nGo OS/Arch: %v/%v\nBuilt at: %v",
+		Version, Commit, GoVersion, runtime.GOOS, runtime.GOARCH, Date,
+	)
+}
